requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-bootstrap-modal-forms==3.0.4
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
filelock==3.13.1
gTTS==2.4.0
gunicorn==21.2.0
idna==3.4
packaging==23.2
platformdirs==3.11.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.4
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
whitenoise==6.6.0
